Login Admin/Uses System.
This project is a simple log-in system which has an admin who can delete or create users, roles and permissions.
And can assign roles and permissions to users. Users can register to the system with a username, password and other information. Also, they can create posts and delete them and edit their profiles.